+++ tests/functional/test_product_endpoints.py
@@ -3,7 +3,7 @@ from contextlib import contextmanager
 from datetime import datetime, timezone
 from unittest.mock import patch, MagicMock, AsyncMock
 
-from app.api.deps import get_current_org_membership
+from app.orgs.deps import get_current_org_membership
 from app.core.database import get_db
 from app.models.org_membership import OrgMembership
 from app.models.product import Product
